test(function): exclude JSON allocation check under race (#25915)
### Root cause

`TestJsonMergePatchWrapperAllocationsDoNotScaleWithRows` measures the complete 8192-row JSON merge path with `testing.AllocsPerRun(3)`, even though its asserted invariant is wrapper allocation behavior. Under `-race`, that path also executes instrumented and race-build-only vector operations, so process-wide allocation deltas outside the intended invariant can be charged to the test.

The same package source passed the Ubuntu/x86 race UT in #25728, then failed an unrelated PR in #25911 with `one=1, many=45`. #25911 only changes the HAKeeper client and its tests, which run in a separate package test binary.

### Fix

Move the allocation-statistics-only test unchanged into a `!race` test file.

This keeps:

- the 1-row versus 8192-row allocation regression assertion in non-race builds;
- all JSON merge functional tests in both race and non-race builds;
- production code and JSON merge behavior unchanged.

### Validation

```text
.agents/skills/mo-dev/scripts/mo-cgo-test -v -count=100 -timeout=10m \
-run '^TestJsonMergePatchWrapperAllocationsDoNotScaleWithRows$' \
./pkg/sql/plan/function

.agents/skills/mo-dev/scripts/mo-cgo-test -race -count=20 -timeout=10m \
-run '^TestJsonMerge($|DepthValidation$)' \
./pkg/sql/plan/function

.agents/skills/mo-dev/scripts/mo-cgo-test -count=20 -timeout=10m \
-run '^TestJsonMerge($|DepthValidation$|PatchWrapperAllocationsDoNotScaleWithRows$)' \
./pkg/sql/plan/function

.agents/skills/mo-dev/scripts/mo-cgo-test -count=1 -timeout=20m \
./pkg/sql/plan/function

.agents/skills/mo-dev/scripts/mo-cgo-test -race -count=1 -timeout=30m \
./pkg/sql/plan/function
```

The race-built test list was also checked to confirm that only the allocation-statistics test is excluded.

Approved by: @XuPeng-SH
